Xavier's Game Update 7/10/25
-Melee attack works fine now. (Thanks Calvin :D)
-Still working on Dodge Roll

+++ Project Acsension/Assets/Scripts/Enemy/EnemyHealth.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public int currentHealth;
-    public int maxHealth = 5;
+    public float currentHealth;
+    public float maxHealth = 5;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
